move tests to integration level
